+++ src/lib/renderer/render.ts
@@ -77,6 +77,7 @@ import { toDegrees } from "../parser/utils/angle.ts";
 import { LineMap as LinesMap } from "../parser/linesmap.ts";
 import { dirname } from "../fs/resolve.ts";
 import { SourceFile } from "../parser/source.ts";
+import { cloneNode } from "../ast/clone.ts";
 
 /**
  * render ast
@@ -125,12 +126,15 @@ export function doRender(
     };
 
     if (options.withParents) {
-        // @ts-ignore
-        let parent: AstNode = data[PARENT];
+        let parent: AstNode;
 
         while (data[PARENT] != null) {
             // @ts-ignore
-            parent = { ...data[PARENT], chi: [{ ...data }] };
+            // parent = { ...data[PARENT], chi: [data] };
+
+            parent = cloneNode(data[PARENT]!) as AstNode;
+            // @ts-ignore
+            parent.chi = [data];
 
             // @ts-ignore
             parent[PARENT] = data[PARENT][PARENT];
@@ -1799,6 +1803,14 @@ export function renderValue(
             return (<FractionToken>(token as NumberToken).val).typ == EnumToken.FractionTokenType
                 ? renderValue(<FractionToken>(token as NumberToken).val, options, cache)
                 : minifyNumber((token as NumberToken).val as number);
+
+        case EnumToken.InfinityTokenType:
+            return "0/0";
+        case EnumToken.NegativeInfinityTokenType:
+            return "-0/0";
+
+            case EnumToken.NaNTokenType:
+                return "NaN";
 
         case EnumToken.AtRuleTokenType:
             return "@" + (token as AtRuleToken).nam;