+++ CHANGELOG.md
@@ -2,9 +2,17 @@
 
 # v1.6.2
 
+## Fix
 - [x] fix base64 encoding error when inline sourcemap is enabled
+
+## Improvements
 - [x] handle division by zero in calc()
 - [x] strip escaped new line from string
+- [x] range media query : parse more expression as range query
+- [x] handle NaN and Infinity in computation
+- [x] parse \<url-modifiers\>
+
+strip escaped new line from string
 ```css
 a[title="a not s\
 o very long title"] {
@@ -19,7 +27,7 @@ a[title="a not so very long title"] {
  color: blue
 }
 ```
-- [x] range media query : parse more expression as range query
+parsing range query is improved
 
 ```css
 
@@ -30,7 +38,43 @@ is parsed as
 ```css
 @media (resolution>=2x) {}
 ```
-- [x] parse \<url-modifiers\>
+
+NaN and Infinity are now handled in calc().
+
+```css
+
+    @media (min-width: 300px) and (max-width: 768px){
+    .s {
+  
+    top: calc(NaN * 1px);
+    line-height: calc(NaN * 1);
+  color: alpha(from red/calc(0/0));
+  background: alpha(from red/calc(pi/0));
+    height: calc(1px * (NaN + 1));  
+    width: calc(NaN + 1 + calc(NaN - 1 + calc(NaN / 1 + calc(NaN * 1))));
+    margin-left: calc(infinity + 1 + calc(infinity - 1 + calc(infinity / 1 + calc(infinity * 1))));
+    margin-right: calc(infinity + 1px + calc(infinity - 1 + calc(infinity / 1 + calc(infinity * 1))));
+
+}
+
+```
+is parsed as 
+
+```css
+@media (300px<=width<=768px) {
+ .s {
+  top: calc(NaN*1px);
+  line-height: calc(NaN);
+  color: alpha(from red/calc(NaN));
+  background: alpha(from red/calc(1/0));
+  height: calc(1px*NaN);
+  width: calc(NaN);
+  margin-left: calc(0/0);
+  margin-right: calc(0/0 + 1px)
+ }
+}
+```
+Parsing url() \<url-modifiers\> is supported.
 
 ```css
 
@@ -48,6 +92,7 @@ is parsed as
 
 # v1.6.1
 
+## Fixes
 - [x] fix bugs when computing math functions.
 - [x] reduce array mutations
 - [x] supporting additional shorthands: scroll-timeline, scroll-padding, scroll-margin.
@@ -74,9 +119,8 @@ a:hover {
 
 # v1.6.0
 
-- [x] added support for math function `tan()`.
-
 ## Improvements
+- [x] added support for math function `tan()`.
 - [x] faster tokenizer
 - [x] ensure transform: rotate(360deg) is not minified to transform: none
 - [x] support input sourcemap from inline sourcemap file. This is only supported by the async parser.